Initial Commit
Adds the vfRemoting.js
Contains two constructors, RemoteAction and RemotePromise.
RemoteAction is the class to handle the remote action sending. It
exposes methods to add callbacks to be fired before the request, on
success, on failure, and on completion (regardless of success /
failure). It also has methods to remove functions.
RemotePromise allows the send function to return a promise. The promise
will resolve during the RemoteAction's onSuccess calls, and reject on
the onFailure calls. It has a getter to expose the RemoteAction.