Makes Promise Safe
Makes the promise function safe for IE users. In the config section, set
if this is promisable (for example, use modernizr.promise). This will
mimic the promise call by setting the provided functions in your then
call into the onSuccess and onFailure.  Returns itself on catch to
prevent exceptions. Sorry no support for catch function beyond this. IE
sucks